refactored helpers into files they are used only once in.
broke the color input test suite into sub suites
